refactor: 사용하지 않는 메서드 및 개행 정리
