+++ src/main/java/com/dinosaur/foodbowl/domain/follow/dao/FollowRepository.java
@@ -10,7 +10,5 @@ public interface FollowRepository extends JpaRepository<Follow, Long> {
 
   long countByFollowing(User following);
 
-  void deleteFollowByFollowerAndFollowing(User follower, User following);
-
   boolean existsByFollowerAndFollowing(User follower, User following);
 }
